requirements.txt
asgiref
Django
django-jalali-date
django-render-partial
jdatetime
Pillow
pytz
sorl-thumbnail
sqlparse
typing-extensions
